requirements.txt
streamlit==1.40.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.0.1
